requirements.txt
langchain
IPython
langchain_community
langgraph
duckduckgo-search
langchain-ollama
pandas
plotly_express
nbformat